Noty - an iOS8 iCloud demo app.
This app is a simple note app.
It is not at all well designed, but it has every functionality to
demonstrate how to use iCloud. How to create, delete, (though no
rename), conflict resolution, etc.
It is created based on a wonderful tutorial on raywenderlich.com:
http://www.raywenderlich.com/6015/beginning-icloud-in-ios-5-tutorial-par
t-1.
The original tutorial was written a few years ago, so some minor
adjustment was needed. I hope beginners of iCloud programming can use
this as a starting point and create some awesome stuff.